# admin/openapi.json
{
  "components": {
    "schemas": {
      "AclCacheView": {
        "properties": {
          "errors": {
            "title": "Errors",
            "type": "integer"
          },
          "evictions": {
            "title": "Evictions",
            "type": "integer"
          },
          "hits": {
            "title": "Hits",
            "type": "integer"
          },
          "misses": {
            "title": "Misses",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "ttl_seconds": {
            "title": "Ttl Seconds",
            "type": "number"
          }
        },
        "required": [
          "ttl_seconds",
          "hits",
          "misses",
          "errors",
          "evictions",
          "size"
        ],
        "title": "AclCacheView",
        "type": "object"
      },
      "AclInvalidate": {
        "properties": {
          "all": {
            "default": false,
            "title": "All",
            "type": "boolean"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          },
          "principal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Principal"
          }
        },
        "title": "AclInvalidate",
        "type": "object"
      },
      "ConnectorCreate": {
        "properties": {
          "kind": {
            "enum": [
              "filesystem",
              "gdrive"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root"
          }
        },
        "required": [
          "name",
          "kind"
        ],
        "title": "ConnectorCreate",
        "type": "object"
      },
      "ConnectorView": {
        "properties": {
          "kind": {
            "enum": [
              "filesystem",
              "gdrive"
            ],
            "title": "Kind",
            "type": "string"
          },
          "last_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IngestRunView"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "needs_request_token": {
            "title": "Needs Request Token",
            "type": "boolean"
          },
          "root": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root"
          }
        },
        "required": [
          "name",
          "kind",
          "root",
          "needs_request_token",
          "last_run"
        ],
        "title": "ConnectorView",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Health": {
        "properties": {
          "chunks": {
            "title": "Chunks",
            "type": "integer"
          },
          "documents": {
            "title": "Documents",
            "type": "integer"
          },
          "embedder": {
            "title": "Embedder",
            "type": "string"
          },
          "index_size": {
            "title": "Index Size",
            "type": "integer"
          },
          "reranker": {
            "title": "Reranker",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "store": {
            "title": "Store",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "status",
          "version",
          "documents",
          "chunks",
          "index_size",
          "embedder",
          "store",
          "reranker"
        ],
        "title": "Health",
        "type": "object"
      },
      "IndexHealth": {
        "properties": {
          "acl_cache": {
            "$ref": "#/components/schemas/AclCacheView"
          },
          "acl_principals": {
            "title": "Acl Principals",
            "type": "integer"
          },
          "bytes_on_disk": {
            "title": "Bytes On Disk",
            "type": "integer"
          },
          "bytes_per_chunk": {
            "title": "Bytes Per Chunk",
            "type": "number"
          },
          "chunks": {
            "title": "Chunks",
            "type": "integer"
          },
          "dimension": {
            "title": "Dimension",
            "type": "integer"
          },
          "documents": {
            "title": "Documents",
            "type": "integer"
          },
          "files": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Files",
            "type": "object"
          },
          "quantizer_fitted": {
            "title": "Quantizer Fitted",
            "type": "boolean"
          },
          "rescore_multiplier": {
            "title": "Rescore Multiplier",
            "type": "integer"
          },
          "rows": {
            "title": "Rows",
            "type": "integer"
          },
          "staged": {
            "title": "Staged",
            "type": "integer"
          }
        },
        "required": [
          "rows",
          "staged",
          "dimension",
          "quantizer_fitted",
          "rescore_multiplier",
          "acl_principals",
          "files",
          "bytes_on_disk",
          "bytes_per_chunk",
          "documents",
          "chunks",
          "acl_cache"
        ],
        "title": "IndexHealth",
        "type": "object"
      },
      "IngestRunView": {
        "properties": {
          "chunks": {
            "title": "Chunks",
            "type": "integer"
          },
          "documents": {
            "title": "Documents",
            "type": "integer"
          },
          "duration_seconds": {
            "title": "Duration Seconds",
            "type": "number"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "skipped_unchanged": {
            "title": "Skipped Unchanged",
            "type": "integer"
          },
          "started_at": {
            "title": "Started At",
            "type": "string"
          },
          "unreadable": {
            "title": "Unreadable",
            "type": "integer"
          }
        },
        "required": [
          "started_at",
          "duration_seconds",
          "documents",
          "chunks",
          "skipped_unchanged",
          "unreadable",
          "error"
        ],
        "title": "IngestRunView",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Perimeter",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/api/acl/invalidate": {
      "post": {
        "description": "The ADR-004 invalidation hook. Call on any membership change, either direction.",
        "operationId": "invalidateAcl",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AclInvalidate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Invalidateacl",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Acl Invalidate",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/api/connectors": {
      "get": {
        "operationId": "listConnectors",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectorView"
                  },
                  "title": "Response Listconnectors",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Connectors",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "createConnector",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectorCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Connector",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/api/connectors/{name}": {
      "delete": {
        "operationId": "deleteConnector",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Connector",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/api/connectors/{name}/ingest": {
      "post": {
        "operationId": "ingestConnector",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestRunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Connector",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/api/index": {
      "get": {
        "operationId": "indexHealth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexHealth"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Index Health",
        "tags": [
          "admin"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "health",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Health"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "ops"
        ]
      }
    }
  }
}